Website/requirements.txt

# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
alembic==1.7.4
argon2-cffi==20.1.0
attrs==21.2.0
backcall==0.2.0
beautifulsoup4==4.10.0
bidict==0.21.3
bleach==4.1.0
bokeh==2.4.0
brotli==1.0.9
# ca-certificates==2020.10.14
cachetools==4.2.4
certifi==2021.5.30
cffi==1.14.6
chardet==3.0.4
charset-normalizer==2.0.6
click==8.0.1
cloudpickle==2.0.0
colorama==0.4.4
colorcet==2.0.6
cycler==0.10.0
dash==2.0.0
dash-core-components==2.0.0
dash-cytoscape==0.3.0
dash-html-components==2.0.0
dash-table==5.0.0
dask==2021.9.1
datashader==0.13.0
datashape==0.5.2
debugpy==1.4.1
decorator==5.0.9
defusedxml==0.7.1
demos==0.0.1
distributed==2021.9.1
dnspython==2.1.0
email-validator==1.1.3
entrypoints==0.3
face-recognition-models==0.3.0
flask==2.0.1
flask-compress==1.10.1
flask-login==0.5.0
flask-migrate==3.1.0
flask-socketio==5.1.1
flask-sqlalchemy==2.5.1
flask-wtf==0.15.1
fsspec==2021.10.0
google==3.0.0
google-api-core==2.1.1
google-api-python-client==2.26.1
google-auth==2.3.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.6
googleapis-common-protos==1.53.0
greenlet==1.1.2
heapdict==1.0.1
holoviews==1.14.6
httplib2==0.20.1
idna==2.6
imageio==2.9.0
ipykernel==6.2.0
ipython==7.27.0
ipython_genutils==0.2.0
ipywidgets==7.6.3
itsdangerous==2.0.1
jedi==0.18.0
jinja2==3.0.1
joblib==1.0.1
jsonschema==3.2.0
kiwisolver==1.3.2
llvmlite==0.37.0
locket==0.2.1
mako==1.1.5
markdown==3.3.4
markupsafe==2.0.1
matplotlib==3.4.3
matplotlib-inline==0.1.2
mistune==0.8.4
msgpack==1.0.2
multipledispatch==0.6.0
nbclient==0.5.4
nbconvert==6.1.0
nbformat==5.1.3
nest-asyncio==1.5.1
networkx==2.6.2
notebook==6.4.3
numba==0.54.0
numpy==1.20.3
oauth2client==4.1.3
oauthlib==3.1.1
#openssl==1.1.1k
packaging==21.0
pandas==1.3.2
pandasql==0.7.3
pandocfilters==1.4.3
panel==0.12.4
param==1.11.1
parso==0.8.2
partd==1.2.0
pickleshare==0.7.5
pillow==8.3.2
pip==21.2.4
plotly==5.3.1
prometheus-client==0.11.0
prompt-toolkit==3.0.20
protobuf==3.18.1
psutil==5.8.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pyct==0.4.8
pygments==2.10.0
pyparsing==2.4.7
pyrsistent==0.18.0
python-dateutil==2.8.2
python-decouple==3.5
python-engineio==4.2.1
python-socketio==5.4.1
pytz==2021.1
pyviz-comms==2.1.0
pywavelets==1.1.1
pyyaml==5.4.1
pyzmq==22.2.1
requests==2.18.4
requests-oauthlib==1.3.0
rfc3339==6.2
rsa==4.7.2
scikit-image==0.18.3
scikit-learn==0.24.2
scipy==1.7.1
seaborn==0.11.2
send2trash==1.8.0
setuptools==52.0.0
simplejson==3.17.5
six==1.16.0
sortedcontainers==2.4.0
soupsieve==2.2.1
sqlalchemy==1.4.25
tblib==1.7.0
tenacity==8.0.1
terminado==0.11.1
testpath==0.5.0
threadpoolctl==2.2.0
tifffile==2021.8.30
toolz==0.11.1
tornado==6.1
tqdm==4.62.3
traitlets==5.0.5
tzdata
uritemplate==4.1.1
urllib3==1.22
vc
wcwidth==0.2.5
webencodings==0.5.1
werkzeug==2.0.1
wheel==0.37.0
widgetsnbextension==3.5.1
wincertstore==0.2
wtforms==2.3.3
xarray==0.19.0
zict==2.0.0
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
# alembic==1.7.4
# argon2-cffi==20.1.0
# attrs==21.2.0
# backcall==0.2.0
# beautifulsoup4==4.10.0
# bidict==0.21.3
# bleach==4.1.0
# bokeh==2.4.0
# brotli==1.0.9
# #ca-certificates==2020.10.14
# cachetools==4.2.4
# certifi==2021.5.3
# cffi==1.14.6
# chardet==3.0.4
# charset-normalizer==2.0.6
# click==8.0.1
# cloudpickle==2.0.0
# colorama==0.4.4
# colorcet==2.0.6
# cycler==0.10.0
# dash==2.0.0
# dash-core-components==2.0.0
# dash-cytoscape==0.3.0
# dash-html-components==2.0.0
# dash-table==5.0.0
# dask==2021.9.1
# datashader==0.13.0
# datashape==0.5.2
# debugpy==1.4.1
# decorator==5.0.9
# defusedxml==0.7.1
# demos==0.0.1
# distributed==2021.9.1
# dnspython==2.1.0
# email-validator==1.1.3
# entrypoints==0.3
# face-recognition-models==0.3.0
# flask==2.0.1
# flask-compress==1.10.1
# flask-login==0.5.0
# flask-migrate==3.1.0
# flask-socketio==5.1.1
# flask-sqlalchemy==2.5.1
# flask-wtf==0.15.1
# fsspec==2021.10.0
# google==3.0.0
# google-api-core==2.1.1
# google-api-python-client==2.26.1
# google-auth==2.3.0
# google-auth-httplib2==0.1.0
# google-auth-oauthlib==0.4.6
# googleapis-common-protos==1.53.0
# greenlet==1.1.2
# heapdict==1.0.1
# holoviews==1.14.6
# httplib2==0.20.1
# idna==2.6
# imageio==2.9.0
# ipykernel==6.2.0
# ipython==7.27.0
# ipython_genutils==0.2.0
# ipywidgets==7.6.3
# itsdangerous==2.0.1
# jedi==0.18.0
# jinja2==3.0.1
# joblib==1.0.1
# jsonschema==3.2.0
# kiwisolver==1.3.2
# llvmlite==0.37.0
# locket==0.2.1
# mako==1.1.5
# markdown==3.3.4
# markupsafe==2.0.1
# matplotlib==3.4.3
# matplotlib-inline==0.1.2
# mistune==0.8.4
# msgpack==1.0.2
# multipledispatch==0.6.0
# nbclient==0.5.4
# nbconvert==6.1.0
# nbformat==5.1.3
# nest-asyncio==1.5.1
# networkx==2.6.2
# notebook==6.4.3
# numba==0.54.0
# numpy==1.20.3
# oauth2client==4.1.3
# oauthlib==3.1.1
# #openssl==1.1.1k
# packaging==21.0
# pandas==1.3.2
# pandasql==0.7.3
# pandocfilters==1.4.3
# panel==0.12.4
# param==1.11.1
# parso==0.8.2
# partd==1.2.0
# pickleshare==0.7.5
# pillow==8.3.2
# pip==21.2.4
# plotly==5.3.1
# prometheus-client==0.11.0
# prompt-toolkit==3.0.20
# protobuf==3.18.1
# psutil==5.8.0
# pyasn1==0.4.8
# pyasn1-modules==0.2.8
# pycparser==2.20
# pyct==0.4.8
# pygments==2.10.0
# pyparsing==2.4.7
# pyrsistent==0.18.0
# python==3.9.6
# python-dateutil==2.8.2
# python-decouple==3.5
# python-engineio==4.2.1
# python-socketio==5.4.1
# pytz==2021.1
# pyviz-comms==2.1.0
# pywavelets==1.1.1
# pywin32==301
# pywinpty==1.1.3
# pyyaml==5.4.1
# pyzmq==22.2.1
# qtconsole==5.1.1
# qtpy==1.10.0
# requests==2.18.4
# requests-oauthlib==1.3.0
# rfc3339==6.2
# rsa==4.7.2
# scikit-image==0.18.3
# scikit-learn==0.24.2
# scipy==1.7.1
# seaborn==0.11.2
# send2trash==1.8.0
# setuptools==52.0.0
# simplejson==3.17.5
# six==1.16.0
# sklearn==0.0
# sortedcontainers==2.4.0
# soupsieve==2.2.1
# sqlalchemy==1.4.25
# sqlite==3.36.0
# tblib==1.7.0
# tenacity==8.0.1
# terminado==0.11.1
# testpath==0.5.0
# threadpoolctl==2.2.0
# tifffile==2021.8.30
# toolz==0.11.1
# tornado==6.1
# tqdm==4.62.3
# traitlets==5.0.5
# typing-extensions==3.10.0.2
# tzdata==2021a
# uritemplate==4.1.1
# urllib3==1.22
# vc==14.2==h21ff451_1
# vs2015_runtime==14.27.29016
# wcwidth==0.2.5
# webencodings==0.5.1
# werkzeug==2.0.1
# wheel==0.37.0
# widgetsnbextension==3.5.1
# wincertstore==0.2
# wtforms==2.3.3
# xarray==0.19.0
# zict==2.0.0

